requirements.txt
librosa==0.9.2
tensorflow-cpu==2.11.0
nltk==3.8.1
estnltk==1.6.9.1b0
pika==1.3.1
torch==1.13.1
torchvision
torchaudio
pyyaml==6.0
pydantic==1.10.4
python-dotenv==0.21.0
ruamel.yaml==0.17.21
phonemizer==3.2.1